refactor: move component fetch specs and row builders to cli/commands/discovery.py
Moves _ComponentFetchSpec, the four fetch-spec constants, all component
fetch/count helpers, and all display-row builders from generator.py to
cli/commands/discovery.py. Re-exports them from generator.py for backwards
compat. Updates test_discovery_component_consistency.py to patch the new
canonical location of make_api_call_with_retry.
